update `config.php` to utilize environment variables
assets are typically overwritten every time the package is updated, so users should have an alternative way to set their bugsnag configuration that prevents having to redo it on updates.  this update allows users to set configuration with their environment variables, but falls back on the sensible defaults. this change is non-breaking.

also use short array syntax